Rename `samples_rate.rs` to `sample_rate.rs` to follow `cpal` 0.7 convention
